% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mkb_aftrek.R
\name{add_mkb_aftrek}
\alias{add_mkb_aftrek}
\title{add_mkb_aftrek}
\usage{
add_mkb_aftrek(df)
}
\arguments{
\item{df}{tibble to add mkb_aftrek to}
}
\value{
df with extra column
}
\description{
add_mkb_aftrek
}
