% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tax_variable.R
\name{add_tax_variable}
\alias{add_tax_variable}
\title{add_tax_variable}
\usage{
add_tax_variable(
  df,
  variable_name = "starters_aftrek",
  needs_true = NA,
  input_col = "bruto"
)
}
\arguments{
\item{df}{tibble to add variable to}

\item{variable_name}{name of the variable to add}

\item{needs_true}{only add variable when this column in df is true}

\item{input_col}{calculate variable based on this column}
}
\value{
df with extra column
}
\description{
add_tax_variable
}
\examples{

df <- tibble::tibble(
  bruto = seq(from = -10000, to = 100000, length.out = 11),
  year = c(2018, 2019, 2020, 2018, 2019, 2020, 2018, 2019, 2020, 2018, 2020),
  ondernemer = c(TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, TRUE),
  starter = c(rep(TRUE, 5), rep(FALSE, 6)),
  oudedag = c(rep(FALSE, 5), rep(TRUE, 6)),
  overige_aftrek = 1500
)

df <- df \%>\%
  add_tax_variable("starters_aftrek", "starter") \%>\%
  add_tax_variable("oudedag_reserve", "oudedag") \%>\%
  add_tax_variable("zelf_aftrek", "ondernemer") \%>\%
  add_mkb_aftrek() \%>\%
  add_winst() \%>\%
  add_tax_variable("alg_hef_korting", input_col = "winst") \%>\%
  add_tax_variable("arb_korting", input_col = "winst") \%>\%
  add_tax_variable("zvw_bijdrage", input_col = "winst") \%>\%
  add_tax_variable("income_tax", input_col = "winst") \%>\%
  add_tariefsaanpassing_extra() \%>\%
  add_total_tax()

}
