% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_total_tax.R
\name{add_total_tax}
\alias{add_total_tax}
\title{add_total_tax}
\usage{
add_total_tax(df)
}
\arguments{
\item{df}{tibble to add column to}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
add_total_tax
}
