% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntervals.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculate Theoretical Confidence Interval}
\usage{
calc_ci(r, alpha = 0.05, na.rm = F)
}
\arguments{
\item{r}{Vector of numeric Values}

\item{alpha}{Numeric value indicates confidence level, i.e. the interval covers \code{1 - alpha}.}

\item{na.rm}{Logical option indicating whethermissing values should be removed.}
}
\value{
List with four numeric vectors: the mean and median values and their respective confidence intervals.
}
\description{
Computes confidence interval calculations for the mean and median. Mean interval based on T test; median interval
based on binomial distribution.
}
