% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_disks.R
\name{calc_disks}
\alias{calc_disks}
\title{Calculate total given some disks and a value}
\usage{
calc_disks(values, disks)
}
\arguments{
\item{values}{numeric vector containing values}

\item{disks}{list of disks. Every disk has a \code{max} and a \code{perc}}
}
\value{
numeric vector containing results per value
}
\description{
Calculate total given some disks and a value
}
\examples{
values <- seq(from = 0, to = 100000, length.out = 11)
disks <- list(
  list(upper_border = 30000, fixed = 3000),
  list(upper_border = 60000, perc = 0.2),
  list(upper_border = Inf, perc = 0.3)
)
calc_disks(values, disks)
}
