% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{cod}
\alias{cod}
\title{Calculate Coefficient of Dispersion (COD)}
\usage{
cod(r, na.rm = F)
}
\arguments{
\item{r}{A numeric vector of values. In the ratio study context, this would be a vector of Assessed-to-Sale Ratio (ASR) values.}

\item{na.rm}{Logical value indicating whether NA values should be removed.}
}
\value{
Numeric vector of length one, indicating the COD of \code{r}.
}
\description{
COD is defined as the average percent deviation from the median. In the tax assessment context, the
COD is a standard metric to assess "horizontal equity", or whether properties of the same market value are assessed
at similar levels.
}
\examples{

cod(runif(10))

}
