% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dateFormat}
\alias{dateFormat}
\title{Reporting format for Dates}
\usage{
dateFormat(d)
}
\arguments{
\item{d}{Vector of either date or character type, formatted "%Y-%m-%d".}
}
\value{
Character vector of formatted dates.
}
\description{
Formats dates in the form: "%m/%d/%Y".
}
