% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioStats.R
\name{formatRatioStats}
\alias{formatRatioStats}
\title{Format Ratio Study Output}
\usage{
formatRatioStats(tab)
}
\arguments{
\item{tab}{Dataframe, output of \cod{ratioStats}}
}
\value{
Dataframe of formatted ratio study statistics
}
\description{
Format Ratio Study Output
}
