% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlierDetection.R
\name{isOutlier}
\alias{isOutlier}
\alias{iqrOutlier}
\alias{ptileOutlier}
\title{Test for Outliers}
\usage{
isOutlier(x, method = "iqr", na.rm = F, ...)

iqrOutlier(x, val = 3, na.rm = F, ...)

ptileOutlier(x, inner_perc = 0.995, na.rm = F, ...)
}
\arguments{
\item{x}{Numeric vector of values}

\item{method}{String indicating method of outlier detection. Values supported are \code{iqr} and \code{ptile}.}

\item{na.rm}{Logical value indicating whether NA values should be removed.}

\item{val}{Numeric value which is multiplied by IQR to identify outlier boundary.}

\item{inner_perc}{Numeric value between 0 and 1 indicating inner percentile of data.}
}
\value{
Logical vector of values
}
\description{
Identify outliers in numeric vector.
}
\details{
Outlier detection methods include inner quartile range test and a simple upper and lower percentile identification.
}
\section{Functions}{
\itemize{
\item \code{iqrOutlier}: Innerquartile range method for identifying outliers.

\item \code{ptileOutlier}: Inner percentile method for identifying outliers.
}}

