% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{prb}
\alias{prb}
\title{Calculate Coefficient of Price-Related Bias (PRB)}
\usage{
prb(assessed, saleprice, ci = 0.95, na.rm = F)
}
\arguments{
\item{assessed}{A numeric vector of assessment values.}

\item{saleprice}{A numeric vector of sale prices.}

\item{ci}{Confidence level for interval.}

\item{na.rm}{Logical value indicating whether NA values should be removed.}
}
\value{
A named numeric vector with the PRB value and high and low bounds.
}
\description{
The coefficient of price-related bias (PRB) is a measure of veritical equity, and is calculated by
regressing the percentage difference from the median ratio on percentage differences in value.
}
\details{
The PRB is a standard metric to assess "vertical equity", or whether high and low value properties
are assessed at similar levels. It is obtained by regressing. Negative values are
indicative of assessment \emph{regressivity}, while positive values are indicative of assessment \emph{progressivity}.
The PRB is generally preferred to the PRD, since the latter is known to be upwardly biased.
}
