% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{prd}
\alias{prd}
\title{Calculate Price-Related Differential}
\usage{
prd(assessed, saleprice, na.rm = FALSE)
}
\arguments{
\item{assessed}{A numeric vector of assessment values.}

\item{saleprice}{A numeric vector of sale prices.}

\item{na.rm}{Logical value indicating whether NA values should be removed.}
}
\value{
Numeric vector of length one.
}
\description{
The price-related differential (PRD) is a measure of veritical equity, and is the mean divided by
the weighted mean.
}
\details{
The PRD is a measure of vertical equity. It is known to have a slight upward bias. For this reason,
the PRB is a regarded as a more reliable metric. PRD values above 1.03 are indicative of assessment \emph{regressivity},
while values below 0.98 is indicative of assessment \emph{progressivity}.
}
