% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioStats.R
\name{ratioStats}
\alias{ratioStats}
\title{Perform Ratio Study}
\usage{
ratioStats(df, property_class = NULL)
}
\arguments{
\item{df}{Dataframe with standardized names}

\item{property_class}{Class of properties to be analyzed, i.e. analysis title.}
}
\value{
Dataframe of standard ratio study statstics
}
\description{
Perform Ratio Study
}
