% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confIntervals.R
\name{robust_ci}
\alias{robust_ci}
\title{Calculate Bootstrapped Confidence Interval}
\usage{
robust_ci(x, y = NULL, nboot = 100, alpha = 0.05, FUN = NULL,
  na.rm = F)
}
\arguments{
\item{x}{Vector of numeric values.}

\item{y}{(Optional) vector of numeric values.}

\item{nboot}{Number indicating number of bootstrap iterations.}

\item{alpha}{Numeric value indicates confidence level, i.e. the interval covers \code{1 - alpha}.}

\item{FUN}{Function to evaluate.}

\item{na.rm}{Logical indicating whether to drop missing values.}
}
\value{
List with four numeric vectors: the mean and median values and their respective confidence intervals.
}
\description{
Computes confidence interval calculations for the mean and median, both based on bootstrap resampling.
}
