% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{standardize}
\alias{standardize}
\title{Standardize Ratio Study Variables}
\usage{
standardize(df, assessed, saleprice = NULL, uid = NULL,
  saledate = NULL, living_area = NULL, lot_area = NULL, asr = NULL,
  x = NULL, y = NULL, calc_asr = T)
}
\arguments{
\item{df}{Dataframe of values}

\item{assessed}{Character string indicating column in \code{df} that is to be renamed \code{assessed}.}

\item{saleprice}{Character string indicating column in \code{df} that is to be renamed \code{saleprice}.}

\item{uid}{Character string indicating column in \code{df} that is to be renamed \code{uid}, i.e. unique ID.}

\item{saledate}{Character string indicating column in \code{df} that is to be renamed \code{saledate}.}

\item{living_area}{Character string indicating column in \code{df} that is to be renamed \code{living_area}.}

\item{lot_area}{Character string indicating column in \code{df} that is to be renamed \code{lot_area}.}

\item{asr}{Character string indicating column in \code{df} that is to be renamed \code{asr}, i.e. assessed to sale ratio.}

\item{x}{Character string indicating column in \code{df} that is to be renamed \code{x}, i.e. longitude coordinate.}

\item{y}{Character string indicating column in \code{df} that is to be renamed \code{y}, i.e. latitude coordinate.}

\item{calc_asr}{Logical indicating whether to calculate the assessed to value ratio (asr).}
}
\value{
Dataframe with renamed columns
}
\description{
Rename Columns
}
