% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{typeStandardization}
\alias{typeStandardization}
\title{Standardize Ratio Study Variable Types}
\usage{
typeStandardization(df)
}
\arguments{
\item{df}{Dataframe output of \code{standardize}}
}
\description{
Standardize Ratio Study Variable Types
}
