% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{weightedMean}
\alias{weightedMean}
\title{Calcualte Weighted Mean}
\usage{
weightedMean(assessed, saleprice, na.rm = F)
}
\arguments{
\item{assessed}{A numeric vector of assessment values.}

\item{saleprice}{A numeric vector of sale prices.}

\item{na.rm}{Logical value indicating whether NA values should be removed.}
}
\value{
Numeric vector of length one.
}
\description{
The weighted mean, defined as the sum of appraised values divided by the sum of sale prices.
}
