% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tariefsaanpassing_extra.R
\name{add_tariefsaanpassing_extra}
\alias{add_tariefsaanpassing_extra}
\title{add_tariefsaanpassing_extra}
\usage{
add_tariefsaanpassing_extra(df)
}
\arguments{
\item{df}{tibble to add column to}
}
\value{
tibble with extra column
}
\description{
add_tariefsaanpassing_extra
}
